+++ CHANGELOG.md
@@ -38,3 +38,4 @@
 - Improve look of Redstone, Glowstone & Ender Ingot Textures + Fluid jsons
 ## 28-03-2026-1.4.6 Modified by xXjudeXx
 - update version number to 1.4.6
+- fix incorrect Anvil Tiers for Redstone, Glowstone & Ender